Fix : add review's update &delete authorization

+++ src/main/java/com/hanghae/clone_project/service/CommentService.java
@@ -7,6 +7,8 @@ import com.hanghae.clone_project.dto.responseDto.ResponseDto;
 import com.hanghae.clone_project.entity.Comment;
 import com.hanghae.clone_project.entity.Product;
 import com.hanghae.clone_project.entity.User;
+import com.hanghae.clone_project.exception.ErrorCode.CustomErrorCode;
+import com.hanghae.clone_project.exception.Exception.RestApiException;
 import com.hanghae.clone_project.repository.CommentRepository;
 import com.hanghae.clone_project.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
@@ -46,17 +48,26 @@ public class CommentService {
     public ResponseEntity<?> updateComment(CommentRequestDto commentRequestDto, Long id, User user) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("댓글을 찾을 수 없습니다."));
+
+        if(!user.getUsername().equals(comment.getUser().getUsername()))
+            throw new RestApiException(CustomErrorCode.UNAUTHORIZED_TOKEN,"수정권한이 없습니다.");
+
         comment.update(commentRequestDto);
         return new ResponseEntity<>(ResponseDto.success(CommentDto.of(comment)), HttpStatus.OK);
     }
 
     //리뷰 삭제
     @Transactional
-    public ResponseEntity<?> deleteComment(Long id) {
+    public ResponseEntity<?> deleteComment(Long id, User user) {
 
         Comment comment = commentRepository
                 .findById(id).orElseThrow(() -> new IllegalArgumentException("댓글을 찾을 수 없습니다."));
+
+        if(!user.getUsername().equals(comment.getUser().getUsername()))
+            throw new RestApiException(CustomErrorCode.UNAUTHORIZED_TOKEN,"수정권한이 없습니다.");
+
         commentRepository.delete(comment);
+
         return new ResponseEntity<>(ResponseDto.success(null), HttpStatus.OK);
 
     }
